ref: centralize lesson controllers in index file
