+++ backend/src/controllers/index.js
@@ -8,5 +8,6 @@ router.use("/professors", require("./professor.controller"));
 router.use("/password", require("./password.controller"));
 router.use("/mail", require("./mail.controller"));
 router.use("/courses", require("./course.controller"));
+router.use("/lessons", require("./lesson.controller"));
 
 module.exports = router;
